third_party/blink/web_tests/platform/win/external/wpt/clipboard-apis/clipboard-copy-selection-line-break.https_11-20-expected.txt

This is a testharness.js-based test.
[FAIL] test preformatted text with line breaks: "\\r\\r\\r"
  assert_equals: Check result from DataTransfer in paste event expected "First\\r\\r\\rSecond" but got "First\\r\\n\\r\\n\\r\\nSecond"
[FAIL] test preformatted text with line breaks: "\\r\\r\\n"
  assert_equals: Check result from DataTransfer in paste event expected "First\\r\\r\\nSecond" but got "First\\r\\n\\r\\nSecond"
[FAIL] test preformatted text with line breaks: "\\r\\n\\r"
  assert_equals: Check result from DataTransfer in paste event expected "First\\r\\n\\rSecond" but got "First\\r\\n\\r\\nSecond"
Harness: the test ran to completion.
